extract helper for copying files
